created two more slices
