+++ src/App.jsx
@@ -1,6 +1,7 @@
 import "./App.css";
 import { useDispatch, useSelector } from "react-redux";
-import { fetchPopularMovies, fetchTopRatedMovies } from "./redux/moviesSlice";
+import { fetchPopularMovies } from "./redux/moviesSlice";
+import { fetchTopRatedMovies } from "./redux/topRatedMoviesSlice";
 import { useEffect, useState } from "react";
 import Slidebar from "./components/Slidebar";
 import Navbar from "./components/Navbar";
@@ -12,7 +13,7 @@ import { Helmet } from "react-helmet";
 function App() {
   const dispatch = useDispatch();
   const popularMovies = useSelector((state) => state.movies.popularMovies);
-  const topRatedMovies = useSelector((state) => state.movies.topRated);
+  const topRatedMovies = useSelector((state) => state.topRatedMovies.topRated);
   const searchResults = useSelector((state) => state.movies.searchResults);
   const movieStatus = useSelector((state) => state.movies.status);
   const [showSearchResults, setShowSearchResults] = useState(true);
@@ -38,7 +39,7 @@ function App() {
       <Helmet>
         <title>Filmkväll - hitta dom bästa filmerna</title>
         <meta
-          name="desription"
+          name="description"
           content="Upptäck dom populäraste och mest gillade filmerna just nu. Sök efter din favoritfilmer och lägg till dom i din lista."
         />
         <meta
