+++ src/redux/movieDetailsSlice.js
@@ -1,0 +1,42 @@
+import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
+
+const API_KEY = import.meta.env.VITE_TMDB_API_KEY;
+
+export const fetchMovieDetails = createAsyncThunk(
+  "movieDetails/fetchMovieDetails",
+  async (id) => {
+    const response = await fetch(
+      `https://api.themoviedb.org/3/movie/${id}?api_key=${API_KEY}`
+    );
+    if (!response.ok) {
+      throw new Error("Failed to fetch movie details");
+    }
+    const data = await response.json();
+    return data;
+  }
+);
+
+const movieDetailsSlice = createSlice({
+  name: "movieDetails",
+  initialState: {
+    movieDetails: [],
+    status: "idle",
+    error: null,
+  },
+  extraReducers: (builder) => {
+    builder
+      .addCase(fetchMovieDetails.pending, (state) => {
+        state.status = "loading";
+      })
+      .addCase(fetchMovieDetails.fulfilled, (state, action) => {
+        state.status = "succeeded";
+        state.movieDetails = action.payload;
+      })
+      .addCase(fetchMovieDetails.rejected, (state, action) => {
+        state.status = "failed";
+        state.error = action.error.message;
+      });
+  },
+});
+
+export default movieDetailsSlice.reducer;
